Revert &str->CellValue impl
